+++ examples/bpmodules/basic/index.js
@@ -1,13 +1,11 @@
 'use strict'
 
-const debug = require('debug')('bigview')
-const fs = require('fs')
 const MyBigView = require('./MyBigView')
 
 module.exports = function (req, res) {
     
     var bigpipe = new MyBigView(req, res, 'basic/index', { title: "测试" })
-    bigpipe.isDynamicLayout = true
+    
     // bigpipe.mode = 'render'
     bigpipe.add(require('./p1'))
     bigpipe.add(require('./p2'))