+++ examples/bpmodules/basic/p2/index.js
@@ -14,7 +14,7 @@ module.exports = class MyPagelet extends Pagelet {
             }
         }
         this.domid = 'pagelet2'
-        this.tpl = 'p2.html'
+        this.tpl = 'p2'
         this.delay = 2000
     }