+++ history.md
@@ -6,7 +6,7 @@
 	- [x] 使用bluebird作为global.Promise，避免外部错误
 	- [x] 去掉bigpipe.debug = false;使用环境变量（桑世龙）
 	- [x] bigview && biglet conosle.log("BIGVIEW" + err) 增加模块标识，便于日志记录（桑世龙）
-	- [ ] this.tpl = 'tpl/index';干掉，设为默认值（桑世龙）
+	- [x] this.tpl = 'tpl/index';干掉，设为默认值（桑世龙）
 	- [ ] this.root = __dirname;有疑问？暂无解决方案，需要思考（桑世龙）
 	- [ ] biglet // custom error function
 	- [ ] 当没有数据的时候，不显示改模块，目前的做法是 return Promise.reject(); 语义上怪（桑世龙）